+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from "./components/App/App/App";
+import App from "./components/App/App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 
@@ -12,7 +12,6 @@ root.render(
       <App />
     </React.StrictMode>
   </BrowserRouter>
-
 );
 
 // If you want to start measuring performance in your app, pass a function
